move img_plotting.py under image
